# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API with Swagger",
        "description": "RESTFULL API sistem penjualan documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/unit": {
            "get": {
                "tags": [
                    "Unit"
                ],
                "summary": "Service unit",
                "description": "Endpoint get unit",
                "operationId": "7335d87c4a3157fe50761cf91d300722",
                "responses": {
                    "200": {
                        "description": "Data found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Unit"
                ],
                "summary": "Service unit",
                "description": "Endpoint add new unit",
                "operationId": "302327015261cb20d1a667f4b5b4c63e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "unit": {
                                                "type": "string"
                                            },
                                            "desc": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "unit": "kg",
                                    "desc": "Kilogram"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "409": {
                        "description": "Unit already exists"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/unit/{id}": {
            "get": {
                "tags": [
                    "Unit"
                ],
                "summary": "Service unit",
                "description": "Endpoint get unit specified id",
                "operationId": "5c3d4f307f1777efebfcf2772f04cb94",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Unit id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Unit"
                ],
                "summary": "Update existing unit",
                "description": "Returns updated unit data",
                "operationId": "update unit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "unit data id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "unit": {
                                                "type": "string"
                                            },
                                            "desc": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "unit": "kg",
                                    "desc": "Kilogram"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Unit"
                ],
                "summary": "Delete existing unit",
                "description": "Deletes a record and returns no content",
                "operationId": "delete unit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "unit id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}